# openapi-expense-tracker.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Expense Tracker API",
    "description" : "Active profile: DEFAULT",
    "version" : "2.8.4"
  },
  "servers" : [ {
    "url" : "http://10.167.7.78:8081",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
    "/expenses/{userId}/update/{expenseId}" : {
      "put" : {
        "tags" : [ "expense-controller" ],
        "summary" : "Update an expense",
        "description" : "Return the updated Expense",
        "operationId" : "update",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "expenseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseUpdateRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseExpenseResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/incomes" : {
      "get" : {
        "tags" : [ "income-controller" ],
        "summary" : "Find all incomes of a user",
        "description" : "Return a list with all incomes",
        "operationId" : "getIncomes",
        "responses" : {
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListIncomeResponseDTO"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "income-controller" ],
        "summary" : "Create a new Income for a user",
        "description" : "Return the created Income",
        "operationId" : "createIncome",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IncomeRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseIncomeResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/expenses/{userId}/create" : {
      "post" : {
        "tags" : [ "expense-controller" ],
        "summary" : "Create a new expense for a specific category",
        "description" : "Return the created Expense",
        "operationId" : "createExpense",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseExpenseResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/incomes/{userId}/total" : {
      "get" : {
        "tags" : [ "income-controller" ],
        "summary" : "Get the total Income of a user",
        "description" : "Return the total income",
        "operationId" : "getTotalIncome",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTotalIncomeResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/incomes/{userId}/net-income" : {
      "get" : {
        "tags" : [ "income-controller" ],
        "summary" : "Get the Net Income of a User(Total Income - Expenses)",
        "description" : "Return the Net Income",
        "operationId" : "getNetIncome",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTotalIncomeResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/expenses/{userId}" : {
      "get" : {
        "tags" : [ "expense-controller" ],
        "summary" : "Get all expenses of the user",
        "description" : "Return a list of all expenses",
        "operationId" : "getAllExpenses",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseNull"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListExpenseResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ApiResponseNull" : {
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          },
          "payload" : {
            "$ref" : "#/components/schemas/Null"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
      "ErrorDTO" : {
        "type" : "object",
        "properties" : {
          "fieldName" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "Null" : {
        "type" : "object"
      },
      "ExpenseUpdateRequestDTO" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : "string",
            "format" : "date"
          },
          "amount" : {
            "type" : "number"
          },
          "description" : {
            "type" : "string"
          },
          "paymentMethod" : {
            "type" : "string",
            "enum" : [ "CASH", "CARD", "BANK_TRANSFER" ]
          },
          "categoryName" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseExpenseResponseDTO" : {
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          },
          "payload" : {
            "$ref" : "#/components/schemas/ExpenseResponseDTO"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
      "CategoryDTO" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          }
        }
      },
      "ExpenseResponseDTO" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : "string",
            "format" : "date"
          },
          "amount" : {
            "type" : "number"
          },
          "description" : {
            "type" : "string"
          },
          "paymentMethod" : {
            "type" : "string",
            "enum" : [ "CASH", "CARD", "BANK_TRANSFER" ]
          },
          "category" : {
            "$ref" : "#/components/schemas/CategoryDTO"
          }
        }
      },
...
      "IncomeRequestDTO" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : "string",
            "format" : "date"
          },
          "amount" : {
            "type" : "number"
          },
          "source" : {
            "type" : "string"
          },
          "paymentMethod" : {
            "type" : "string",
            "enum" : [ "CASH", "CARD", "BANK_TRANSFER" ]
          }
        },
        "required" : [ "amount", "date", "paymentMethod", "source" ]
      },
      "ApiResponseIncomeResponseDTO" : {
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          },
          "payload" : {
            "$ref" : "#/components/schemas/IncomeResponseDTO"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
      "IncomeResponseDTO" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : "string",
            "format" : "date"
          },
          "amount" : {
            "type" : "number"
          },
          "source" : {
            "type" : "string"
          },
          "paymentMethod" : {
            "type" : "string",
            "enum" : [ "CASH", "CARD", "BANK_TRANSFER" ]
          }
        }
      },
      "ExpenseRequestDTO" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : "string",
            "format" : "date"
          },
          "amount" : {
            "type" : "number"
          },
          "description" : {
            "type" : "string"
          },
          "paymentMethod" : {
            "type" : "string",
            "enum" : [ "CASH", "CARD", "BANK_TRANSFER" ]
          },
          "categoryName" : {
            "type" : "string"
          }
        },
        "required" : [ "amount", "categoryName", "date", "description", "paymentMethod" ]
      },
...
      "ApiResponseListIncomeResponseDTO" : {
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          },
          "payload" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IncomeResponseDTO"
            }
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
      "ApiResponseTotalIncomeResponseDTO" : {
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          },
          "payload" : {
            "$ref" : "#/components/schemas/TotalIncomeResponseDTO"
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
      "TotalIncomeResponseDTO" : {
        "type" : "object",
        "properties" : {
          "totalIncome" : {
            "type" : "number"
          }
        }
      },
      "ApiResponseListExpenseResponseDTO" : {
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          },
          "payload" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExpenseResponseDTO"
            }
          },
          "errors" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDTO"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "name" : "bearerAuth",
        "in" : "header",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}